ExplainabilityMethods

ExplainabilityMethods.EpsilonRule

ExplainabilityMethods.GammaRule

ExplainabilityMethods.IndexNS

ExplainabilityMethods.LRP

ExplainabilityMethods.MaxActivationNS

ExplainabilityMethods.ZBoxRule

ExplainabilityMethods.ZeroRule

ExplainabilityMethods.analyze

ExplainabilityMethods.check_ouput_softmax

ExplainabilityMethods.drop_singleton_dims

ExplainabilityMethods.flatten_chain

ExplainabilityMethods.heatmap

ExplainabilityMethods.modify_denominator

ExplainabilityMethods.modify_layer

ExplainabilityMethods.modify_params

ExplainabilityMethods.safedivide

ExplainabilityMethods.set_weights

ExplainabilityMethods.stabilize_denom

ExplainabilityMethods.strip_softmax

ExplainabilityMethods.EpsilonRule — Type

EpsilonRule(; ϵ=1f-6)

Constructor for LRP-$ϵ$ rule. Commonly used on middle layers.

Arguments:

ϵ: Optional stabilization parameter, defaults to 1f-6.

ExplainabilityMethods.GammaRule — Type

GammaRule(; γ=0.25)

Constructor for LRP-$γ$ rule. Commonly used on lower layers.

Arguments:

γ: Optional multiplier for added positive weights, defaults to 0.25.

ExplainabilityMethods.IndexNS — Type

IndexNS(index)

Neuron selector that picks the output neuron at the given index.

ExplainabilityMethods.LRP — Type

LRP(c::Chain, r::AbstractLRPRule)
LRP(c::Chain, rs::AbstractVector{<:AbstractLRPRule})
LRP(layers::AbstractVector{LRPLayer})

Analyzer that applies LRP.

ExplainabilityMethods.MaxActivationNS — Type

MaxActivationNS()

Neuron selector that picks the output neuron with the highest activation.

ExplainabilityMethods.ZBoxRule — Type

ZBoxRule()

Constructor for LRP-z^{\mathcal{B}}`-rule. Commonly used on the first layer for pixel input.

ExplainabilityMethods.ZeroRule — Type

ZeroRule()

Constructor for LRP-0 rule. Commonly used on upper layers.

ExplainabilityMethods.analyze — Method

analyze(input, method)
analyze(input, method, neuron_selection)

Return raw classifier output and explanation. If neuron_selection is specified, the explanation will be calculated for that neuron. Otherwise, the output neuron with the highest activation is automatically chosen.

ExplainabilityMethods.check_ouput_softmax — Method

check_ouput_softmax(model)

Check whether model has softmax activation on output. Return the model if it doesn't, throw error otherwise.

ExplainabilityMethods.drop_singleton_dims — Method

drop_singleton_dims(a)

Drop dimensions of size 1 from array.

ExplainabilityMethods.flatten_chain — Method

flatten_chain(c)

Flatten a Flux chain containing Flux chains.

ExplainabilityMethods.heatmap — Method

heatmap(expl; kwargs...)

Visualize explanation.

ExplainabilityMethods.modify_denominator — Method

modify_denominator!(d, rule)

Function that modifies zₖ on the forward pass, e.g. for numerical stability.

ExplainabilityMethods.modify_layer — Method

modify_layer(rule, layer)

Applies modify_params to layer if it has parameters

ExplainabilityMethods.modify_params — Method

modify_params!(rule, W, b)

Function that modifies weights and biases before applying relevance propagation.

ExplainabilityMethods.safedivide — Method

safedivide(a, b; eps = 1f-6)

Elementwise division of two matrices avoiding zero terms in the denominator by replacing them with eps.

ExplainabilityMethods.set_weights — Method

set_weights(layer, W, b)

Duplicate layer using weights W, b.

ExplainabilityMethods.stabilize_denom — Method

stabilize_denom(d; eps = 1f-6)

Replace zero terms of a matrix d with eps.

ExplainabilityMethods.strip_softmax — Method

strip_softmax(model)

Remove softmax activation on model output if it exists.