LRP analyzer

Refer to LRP for documentation on the LRP analyzer.

LRP rules

ExplainableAI.ZeroRule — Type

ZeroRule()

LRP-$0$ rule. Commonly used on upper layers.

Definition

Propagates relevance $R^{k+1}$ at layer output to $R^k$ at layer input according to

\[R_j^k = \sum_i \frac{W_{ij}a_j^k}{\sum_l W_{il}a_l^k+b_i} R_i^{k+1}\]

References

S. Bach et al., On Pixel-Wise Explanations for Non-Linear Classifier Decisions by Layer-Wise Relevance Propagation

ExplainableAI.EpsilonRule — Type

EpsilonRule([epsilon=1.0e-6])

LRP-$ϵ$ rule. Commonly used on middle layers.

Definition

Propagates relevance $R^{k+1}$ at layer output to $R^k$ at layer input according to

\[R_j^k = \sum_i\frac{W_{ij}a_j^k}{\epsilon +\sum_{l}W_{il}a_l^k+b_i} R_i^{k+1}\]

Optional arguments

epsilon: Optional stabilization parameter, defaults to 1.0e-6.

References

S. Bach et al., On Pixel-Wise Explanations for Non-Linear Classifier Decisions by Layer-Wise Relevance Propagation

ExplainableAI.GammaRule — Type

GammaRule([gamma=0.25])

LRP-$γ$ rule. Commonly used on lower layers.

Definition

Propagates relevance $R^{k+1}$ at layer output to $R^k$ at layer input according to

\[R_j^k = \sum_i\frac{(W_{ij}+\gamma W_{ij}^+)a_j^k} {\sum_l(W_{il}+\gamma W_{il}^+)a_l^k+(b_i+\gamma b_i^+)} R_i^{k+1}\]

Optional arguments

gamma: Optional multiplier for added positive weights, defaults to 0.25.

References

G. Montavon et al., Layer-Wise Relevance Propagation: An Overview

ExplainableAI.WSquareRule — Type

WSquareRule()

LRP-$w²$ rule. Commonly used on the first layer when values are unbounded.

Definition

Propagates relevance $R^{k+1}$ at layer output to $R^k$ at layer input according to

\[R_j^k = \sum_i\frac{W_{ij}^2}{\sum_l W_{il}^2} R_i^{k+1}\]

References

G. Montavon et al., Explaining Nonlinear Classification Decisions with Deep Taylor Decomposition

ExplainableAI.FlatRule — Type

FlatRule()

LRP-Flat rule. Similar to the WSquareRule, but with all weights set to one and all bias terms set to zero.

Definition

Propagates relevance $R^{k+1}$ at layer output to $R^k$ at layer input according to

\[R_j^k = \sum_i\frac{1}{\sum_l 1} R_i^{k+1} = \sum_i\frac{1}{n_i} R_i^{k+1}\]

where $n_i$ is the number of input neurons connected to the output neuron at index $i$.

References

S. Lapuschkin et al., Unmasking Clever Hans predictors and assessing what machines really learn

ExplainableAI.AlphaBetaRule — Type

AlphaBetaRule([alpha=2.0, beta=1.0])

LRP-$αβ$ rule. Weights positive and negative contributions according to the parameters alpha and beta respectively. The difference $α-β$ must be equal to one. Commonly used on lower layers.

Definition

Propagates relevance $R^{k+1}$ at layer output to $R^k$ at layer input according to

\[R_j^k = \sum_i\left( \alpha\frac{\left(W_{ij}a_j^k\right)^+}{\sum_l\left(W_{il}a_l^k+b_i\right)^+} -\beta\frac{\left(W_{ij}a_j^k\right)^-}{\sum_l\left(W_{il}a_l^k+b_i\right)^-} \right) R_i^{k+1}\]

Optional arguments

alpha: Multiplier for the positive output term, defaults to 2.0.

beta: Multiplier for the negative output term, defaults to 1.0.

References

S. Bach et al., On Pixel-Wise Explanations for Non-Linear Classifier Decisions by Layer-Wise Relevance Propagation

G. Montavon et al., Layer-Wise Relevance Propagation: An Overview

ExplainableAI.ZPlusRule — Type

ZPlusRule()

LRP-$z⁺$ rule. Commonly used on lower layers.

Equivalent to AlphaBetaRule(1.0f0, 0.0f0), but slightly faster. See also AlphaBetaRule.

Definition

Propagates relevance $R^{k+1}$ at layer output to $R^k$ at layer input according to

\[R_j^k = \sum_i\frac{\left(W_{ij}a_j^k\right)^+}{\sum_l\left(W_{il}a_l^k+b_i\right)^+} R_i^{k+1}\]

References

S. Bach et al., On Pixel-Wise Explanations for Non-Linear Classifier Decisions by Layer-Wise Relevance Propagation

G. Montavon et al., Explaining Nonlinear Classification Decisions with Deep Taylor Decomposition

ExplainableAI.ZBoxRule — Type

ZBoxRule(low, high)

LRP-$zᴮ$-rule. Commonly used on the first layer for pixel input.

The parameters low and high should be set to the lower and upper bounds of the input features, e.g. 0.0 and 1.0 for raw image data. It is also possible to provide two arrays of that match the input size.

Definition

Propagates relevance $R^{k+1}$ at layer output to $R^k$ at layer input according to

\[R_j^k=\sum_i \frac{W_{ij}a_j^k - W_{ij}^{+}l_j - W_{ij}^{-}h_j} {\sum_l W_{il}a_l^k+b_i - \left(W_{il}^{+}l_l+b_i^{+}\right) - \left(W_{il}^{-}h_l+b_i^{-}\right)} R_i^{k+1}\]

References

G. Montavon et al., Layer-Wise Relevance Propagation: An Overview

ExplainableAI.PassRule — Type

PassRule()

Pass-through rule. Passes relevance through to the lower layer.

Supports layers with constant input and output shapes, e.g. reshaping layers.

Definition

Propagates relevance $R^{k+1}$ at layer output to $R^k$ at layer input according to

\[R_j^k = R_j^{k+1}\]

ExplainableAI.GeneralizedGammaRule — Type

GeneralizedGammaRule([gamma=0.25])

Generalized LRP-$γ$ rule. Can be used on layers with leakyrelu activation functions.

Definition

Propagates relevance $R^{k+1}$ at layer output to $R^k$ at layer input according to

\[R_j^k = \sum_i\frac {(W_{ij}+\gamma W_{ij}^+)a_j^+ +(W_{ij}+\gamma W_{ij}^-)a_j^-} {\sum_l(W_{il}+\gamma W_{il}^+)a_j^+ +(W_{il}+\gamma W_{il}^-)a_j^- +(b_i+\gamma b_i^+)} I(z_k>0) \cdot R^{k+1}_i +\sum_i\frac {(W_{ij}+\gamma W_{ij}^-)a_j^+ +(W_{ij}+\gamma W_{ij}^+)a_j^-} {\sum_l(W_{il}+\gamma W_{il}^-)a_j^+ +(W_{il}+\gamma W_{il}^+)a_j^- +(b_i+\gamma b_i^-)} I(z_k<0) \cdot R^{k+1}_i\]

Optional arguments

gamma: Optional multiplier for added positive weights, defaults to 0.25.

References

L. Andéol et al., Learning Domain Invariant Representations by Joint Wasserstein Distance Minimization

For manual rule assignment, use ChainTuple and ParallelTuple, matching the model structure:

ExplainableAI.ChainTuple — Type

ChainTuple(xs)

Thin wrapper around Tuple for use with Flux.jl models.

Combining ChainTuple and ParallelTuple, data xs can be stored while preserving the structure of a Flux model without risking type piracy.

ExplainableAI.ParallelTuple — Type

ParallelTuple(xs)

Thin wrapper around Tuple for use with Flux.jl models.

Combining ChainTuple and ParallelTuple, data xs can be stored while preserving the structure of a Flux model without risking type piracy.

Composites

Applying composites

ExplainableAI.Composite — Type

Composite(primitives...)
Composite(default_rule, primitives...)

Automatically contructs a list of LRP-rules by sequentially applying composite primitives.

Primitives

To apply a single rule, use:

LayerMap to apply a rule to the n-th layer of a model

GlobalMap to apply a rule to all layers

RangeMap to apply a rule to a positional range of layers

FirstLayerMap to apply a rule to the first layer

LastLayerMap to apply a rule to the last layer

To apply a set of rules to layers based on their type, use:

GlobalTypeMap to apply a dictionary that maps layer types to LRP-rules

RangeTypeMap for a TypeMap on generalized ranges

FirstLayerTypeMap for a TypeMap on the first layer of a model

LastLayerTypeMap for a TypeMap on the last layer

FirstNTypeMap for a TypeMap on the first n layers

Example

Using a VGG11 model:

julia> composite = Composite(
           GlobalTypeMap(
               ConvLayer => AlphaBetaRule(),
               Dense => EpsilonRule(),
               PoolingLayer => EpsilonRule(),
               DropoutLayer => PassRule(),
               ReshapingLayer => PassRule(),
           ),
           FirstNTypeMap(7, Conv => FlatRule()),
       );

julia> analyzer = LRP(model, composite)
LRP(
  Conv((3, 3), 3 => 64, relu, pad=1)    => FlatRule(),
  MaxPool((2, 2))                       => EpsilonRule{Float32}(1.0f-6),
  Conv((3, 3), 64 => 128, relu, pad=1)  => FlatRule(),
  MaxPool((2, 2))                       => EpsilonRule{Float32}(1.0f-6),
  Conv((3, 3), 128 => 256, relu, pad=1) => FlatRule(),
  Conv((3, 3), 256 => 256, relu, pad=1) => FlatRule(),
  MaxPool((2, 2))                       => EpsilonRule{Float32}(1.0f-6),
  Conv((3, 3), 256 => 512, relu, pad=1) => AlphaBetaRule{Float32}(2.0f0, 1.0f0),
  Conv((3, 3), 512 => 512, relu, pad=1) => AlphaBetaRule{Float32}(2.0f0, 1.0f0),
  MaxPool((2, 2))                       => EpsilonRule{Float32}(1.0f-6),
  Conv((3, 3), 512 => 512, relu, pad=1) => AlphaBetaRule{Float32}(2.0f0, 1.0f0),
  Conv((3, 3), 512 => 512, relu, pad=1) => AlphaBetaRule{Float32}(2.0f0, 1.0f0),
  MaxPool((2, 2))                       => EpsilonRule{Float32}(1.0f-6),
  MLUtils.flatten                       => PassRule(),
  Dense(25088 => 4096, relu)            => EpsilonRule{Float32}(1.0f-6),
  Dropout(0.5)                          => PassRule(),
  Dense(4096 => 4096, relu)             => EpsilonRule{Float32}(1.0f-6),
  Dropout(0.5)                          => PassRule(),
  Dense(4096 => 1000)                   => EpsilonRule{Float32}(1.0f-6),
)

ExplainableAI.lrp_rules — Function

lrp_rules(model, composite)

Apply a composite to obtain LRP-rules for a given Flux model.

Composite primitives

Mapping layers to rules

Composite primitives that apply a single rule:

ExplainableAI.LayerMap — Type

LayerMap(index, rule)

Composite primitive that maps an LRP-rule to all layers in the model at the given index. The index can either be an integer or a tuple of integers to map a rule to a specific layer in nested Flux Chains.

See show_layer_indices to print layer indices and Composite for an example.

ExplainableAI.GlobalMap — Type

GlobalMap(rule)

Composite primitive that maps an LRP-rule to all layers in the model.

See Composite for an example.

ExplainableAI.RangeMap — Type

RangeMap(range, rule)

Composite primitive that maps an LRP-rule to the specified positional range of layers in the model.

See Composite for an example.

ExplainableAI.FirstLayerMap — Type

FirstLayerMap(rule)

Composite primitive that maps an LRP-rule to the first layer in the model.

See Composite for an example.

ExplainableAI.LastLayerMap — Type

LastLayerMap(rule)

Composite primitive that maps an LRP-rule to the last layer in the model.

See Composite for an example.

To apply LayerMap to nested Flux Chains or Parallel layers, make use of show_layer_indices:

ExplainableAI.show_layer_indices — Function

show_layer_indices(model)

Print layer indices of Flux models. This is primarily a utility to help define LayerMap primitives.

Mapping layers to rules based on type

Composite primitives that apply rules based on the layer type:

ExplainableAI.GlobalTypeMap — Type

GlobalTypeMap(map)

Composite primitive that maps layer types to LRP rules based on a list of type-rule-pairs map.

See Composite for an example.

ExplainableAI.RangeTypeMap — Type

RangeTypeMap(range, map)

Composite primitive that maps layer types to LRP rules based on a list of type-rule-pairs map within the specified range of layers in the model.

See Composite for an example.

ExplainableAI.FirstLayerTypeMap — Type

FirstLayerTypeMap(map)

Composite primitive that maps the type of the first layer of the model to LRP rules based on a list of type-rule-pairs map.

See Composite for an example.

ExplainableAI.LastLayerTypeMap — Type

LastLayerTypeMap(map)

Composite primitive that maps the type of the last layer of the model to LRP rules based on a list of type-rule-pairs map.

See Composite for an example.

ExplainableAI.FirstNTypeMap — Type

FirstNTypeMap(n, map)

Composite primitive that maps layer types to LRP rules based on a list of type-rule-pairs map within the first n layers in the model.

See Composite for an example.

Union types for composites

The following exported union types types can be used to define TypeMaps:

ExplainableAI.ConvLayer — Type

Union type for convolutional layers.

ExplainableAI.PoolingLayer — Type

Union type for pooling layers.

ExplainableAI.DropoutLayer — Type

Union type for dropout layers.

ExplainableAI.ReshapingLayer — Type

Union type for reshaping layers such as flatten.

ExplainableAI.NormalizationLayer — Type

Union type for normalization layers.

Composite presets

ExplainableAI.EpsilonGammaBox — Function

EpsilonGammaBox(low, high; [epsilon=1.0f-6, gamma=0.25f0])

Composite using the following primitives:

julia> EpsilonGammaBox(-3.0f0, 3.0f0)
Composite(
  GlobalTypeMap(  # all layers
    Flux.Conv               => ExplainableAI.GammaRule{Float32}(0.25f0),
    Flux.ConvTranspose      => ExplainableAI.GammaRule{Float32}(0.25f0),
    Flux.CrossCor           => ExplainableAI.GammaRule{Float32}(0.25f0),
    Flux.Dense              => ExplainableAI.EpsilonRule{Float32}(1.0f-6),
    typeof(NNlib.dropout)   => ExplainableAI.PassRule(),
    Flux.AlphaDropout       => ExplainableAI.PassRule(),
    Flux.Dropout            => ExplainableAI.PassRule(),
    Flux.BatchNorm          => ExplainableAI.PassRule(),
    typeof(Flux.flatten)    => ExplainableAI.PassRule(),
    typeof(MLUtils.flatten) => ExplainableAI.PassRule(),
    typeof(identity)        => ExplainableAI.PassRule(),
 ),
  FirstLayerTypeMap(  # first layer
    Flux.Conv          => ExplainableAI.ZBoxRule{Float32}(-3.0f0, 3.0f0),
    Flux.ConvTranspose => ExplainableAI.ZBoxRule{Float32}(-3.0f0, 3.0f0),
    Flux.CrossCor      => ExplainableAI.ZBoxRule{Float32}(-3.0f0, 3.0f0),
 ),
)

ExplainableAI.EpsilonPlus — Function

EpsilonPlus(; [epsilon=1.0f-6])

Composite using the following primitives:

julia> EpsilonPlus()
Composite(
  GlobalTypeMap(  # all layers
    Flux.Conv               => ExplainableAI.ZPlusRule(),
    Flux.ConvTranspose      => ExplainableAI.ZPlusRule(),
    Flux.CrossCor           => ExplainableAI.ZPlusRule(),
    Flux.Dense              => ExplainableAI.EpsilonRule{Float32}(1.0f-6),
    typeof(NNlib.dropout)   => ExplainableAI.PassRule(),
    Flux.AlphaDropout       => ExplainableAI.PassRule(),
    Flux.Dropout            => ExplainableAI.PassRule(),
    Flux.BatchNorm          => ExplainableAI.PassRule(),
    typeof(Flux.flatten)    => ExplainableAI.PassRule(),
    typeof(MLUtils.flatten) => ExplainableAI.PassRule(),
    typeof(identity)        => ExplainableAI.PassRule(),
 ),
)

ExplainableAI.EpsilonAlpha2Beta1 — Function

EpsilonAlpha2Beta1(; [epsilon=1.0f-6])

Composite using the following primitives:

julia> EpsilonAlpha2Beta1()
Composite(
  GlobalTypeMap(  # all layers
    Flux.Conv               => ExplainableAI.AlphaBetaRule{Float32}(2.0f0, 1.0f0),
    Flux.ConvTranspose      => ExplainableAI.AlphaBetaRule{Float32}(2.0f0, 1.0f0),
    Flux.CrossCor           => ExplainableAI.AlphaBetaRule{Float32}(2.0f0, 1.0f0),
    Flux.Dense              => ExplainableAI.EpsilonRule{Float32}(1.0f-6),
    typeof(NNlib.dropout)   => ExplainableAI.PassRule(),
    Flux.AlphaDropout       => ExplainableAI.PassRule(),
    Flux.Dropout            => ExplainableAI.PassRule(),
    Flux.BatchNorm          => ExplainableAI.PassRule(),
    typeof(Flux.flatten)    => ExplainableAI.PassRule(),
    typeof(MLUtils.flatten) => ExplainableAI.PassRule(),
    typeof(identity)        => ExplainableAI.PassRule(),
 ),
)

ExplainableAI.EpsilonPlusFlat — Function

EpsilonPlusFlat(; [epsilon=1.0f-6])

Composite using the following primitives:

julia> EpsilonPlusFlat()
Composite(
  GlobalTypeMap(  # all layers
    Flux.Conv               => ExplainableAI.ZPlusRule(),
    Flux.ConvTranspose      => ExplainableAI.ZPlusRule(),
    Flux.CrossCor           => ExplainableAI.ZPlusRule(),
    Flux.Dense              => ExplainableAI.EpsilonRule{Float32}(1.0f-6),
    typeof(NNlib.dropout)   => ExplainableAI.PassRule(),
    Flux.AlphaDropout       => ExplainableAI.PassRule(),
    Flux.Dropout            => ExplainableAI.PassRule(),
    Flux.BatchNorm          => ExplainableAI.PassRule(),
    typeof(Flux.flatten)    => ExplainableAI.PassRule(),
    typeof(MLUtils.flatten) => ExplainableAI.PassRule(),
    typeof(identity)        => ExplainableAI.PassRule(),
 ),
  FirstLayerTypeMap(  # first layer
    Flux.Conv          => ExplainableAI.FlatRule(),
    Flux.ConvTranspose => ExplainableAI.FlatRule(),
    Flux.CrossCor      => ExplainableAI.FlatRule(),
 ),
)

ExplainableAI.EpsilonAlpha2Beta1Flat — Function

EpsilonAlpha2Beta1Flat(; [epsilon=1.0f-6])

Composite using the following primitives:

julia> EpsilonAlpha2Beta1Flat()
Composite(
  GlobalTypeMap(  # all layers
    Flux.Conv               => ExplainableAI.AlphaBetaRule{Float32}(2.0f0, 1.0f0),
    Flux.ConvTranspose      => ExplainableAI.AlphaBetaRule{Float32}(2.0f0, 1.0f0),
    Flux.CrossCor           => ExplainableAI.AlphaBetaRule{Float32}(2.0f0, 1.0f0),
    Flux.Dense              => ExplainableAI.EpsilonRule{Float32}(1.0f-6),
    typeof(NNlib.dropout)   => ExplainableAI.PassRule(),
    Flux.AlphaDropout       => ExplainableAI.PassRule(),
    Flux.Dropout            => ExplainableAI.PassRule(),
    Flux.BatchNorm          => ExplainableAI.PassRule(),
    typeof(Flux.flatten)    => ExplainableAI.PassRule(),
    typeof(MLUtils.flatten) => ExplainableAI.PassRule(),
    typeof(identity)        => ExplainableAI.PassRule(),
 ),
  FirstLayerTypeMap(  # first layer
    Flux.Conv          => ExplainableAI.FlatRule(),
    Flux.ConvTranspose => ExplainableAI.FlatRule(),
    Flux.CrossCor      => ExplainableAI.FlatRule(),
 ),
)

Custom rules

These utilities can be used to define custom rules without writing boilerplate code. To extend these functions, explicitly import them:

ExplainableAI.modify_input — Function

modify_input(rule, input)

Modify input activation before computing relevance propagation.

ExplainableAI.modify_denominator — Function

modify_denominator(rule, d)

Modify denominator $z$ for numerical stability on the forward pass.

ExplainableAI.modify_parameters — Function

modify_parameters(rule, parameter)

Modify parameters before computing the relevance.

Note

Use of a custom function modify_layer will overwrite functionality of modify_parameters, modify_weight and modify_bias for the implemented combination of rule and layer types. This is due to the fact that internally, modify_weight and modify_bias are called by the default implementation of modify_layer. modify_weight and modify_bias in turn call modify_parameters by default.

The default call structure looks as follows:

┌─────────────────────────────────────────┐
│              modify_layer               │
└─────────┬─────────────────────┬─────────┘
          │ calls               │ calls
┌─────────▼─────────┐ ┌─────────▼─────────┐
│   modify_weight   │ │    modify_bias    │
└─────────┬─────────┘ └─────────┬─────────┘
          │ calls               │ calls
┌─────────▼─────────┐ ┌─────────▼─────────┐
│ modify_parameters │ │ modify_parameters │
└───────────────────┘ └───────────────────┘

ExplainableAI.modify_weight — Function

modify_weight(rule, weight)

Modify layer weights before computing the relevance.

Note

Use of a custom function modify_layer will overwrite functionality of modify_parameters, modify_weight and modify_bias for the implemented combination of rule and layer types. This is due to the fact that internally, modify_weight and modify_bias are called by the default implementation of modify_layer. modify_weight and modify_bias in turn call modify_parameters by default.

The default call structure looks as follows:

┌─────────────────────────────────────────┐
│              modify_layer               │
└─────────┬─────────────────────┬─────────┘
          │ calls               │ calls
┌─────────▼─────────┐ ┌─────────▼─────────┐
│   modify_weight   │ │    modify_bias    │
└─────────┬─────────┘ └─────────┬─────────┘
          │ calls               │ calls
┌─────────▼─────────┐ ┌─────────▼─────────┐
│ modify_parameters │ │ modify_parameters │
└───────────────────┘ └───────────────────┘

ExplainableAI.modify_bias — Function

modify_bias(rule, bias)

Modify layer bias before computing the relevance.

Note

Use of a custom function modify_layer will overwrite functionality of modify_parameters, modify_weight and modify_bias for the implemented combination of rule and layer types. This is due to the fact that internally, modify_weight and modify_bias are called by the default implementation of modify_layer. modify_weight and modify_bias in turn call modify_parameters by default.

The default call structure looks as follows:

┌─────────────────────────────────────────┐
│              modify_layer               │
└─────────┬─────────────────────┬─────────┘
          │ calls               │ calls
┌─────────▼─────────┐ ┌─────────▼─────────┐
│   modify_weight   │ │    modify_bias    │
└─────────┬─────────┘ └─────────┬─────────┘
          │ calls               │ calls
┌─────────▼─────────┐ ┌─────────▼─────────┐
│ modify_parameters │ │ modify_parameters │
└───────────────────┘ └───────────────────┘

ExplainableAI.modify_layer — Function

modify_layer(rule, layer)

Modify layer before computing the relevance.

Note

Use of a custom function modify_layer will overwrite functionality of modify_parameters, modify_weight and modify_bias for the implemented combination of rule and layer types. This is due to the fact that internally, modify_weight and modify_bias are called by the default implementation of modify_layer. modify_weight and modify_bias in turn call modify_parameters by default.

The default call structure looks as follows:

┌─────────────────────────────────────────┐
│              modify_layer               │
└─────────┬─────────────────────┬─────────┘
          │ calls               │ calls
┌─────────▼─────────┐ ┌─────────▼─────────┐
│   modify_weight   │ │    modify_bias    │
└─────────┬─────────┘ └─────────┬─────────┘
          │ calls               │ calls
┌─────────▼─────────┐ ┌─────────▼─────────┐
│ modify_parameters │ │ modify_parameters │
└───────────────────┘ └───────────────────┘

ExplainableAI.is_compatible — Function

is_compatible(rule, layer)

Check compatibility of a LRP-Rule with layer type.

Compatibility settings:

ExplainableAI.LRP_CONFIG.supports_layer — Function

LRP_CONFIG.supports_layer(layer)

Check whether LRP can be used on a layer or a Chain. To extend LRP to your own layers, define:

LRP_CONFIG.supports_layer(::MyLayer) = true          # for structs
LRP_CONFIG.supports_layer(::typeof(mylayer)) = true  # for functions

ExplainableAI.LRP_CONFIG.supports_activation — Function

LRP_CONFIG.supports_activation(σ)

Check whether LRP can be used on a given activation function. To extend LRP to your own activation functions, define:

LRP_CONFIG.supports_activation(::typeof(myactivation)) = true  # for functions
LRP_CONFIG.supports_activation(::MyActivation) = true          # for structs

CRP

ExplainableAI.CRP — Type

CRP(lrp_analyzer, layer, concepts)

Use Concept Relevance Propagation to explain the output of a neural network with respect to specific neurons in a given layer.

Arguments

lrp_analyzer::LRP: LRP analyzer

layer::Int: Index of layer after which the concept is located

concepts: Concept to explain.

See also TopNConcepts and IndexedConcepts.

References

[1] R. Achtibat et al., From attribution maps to human-understandable explanations through Concept Relevance Propagation

ExplainableAI.TopNConcepts — Type

TopNConcepts(n)

Select top-n concepts by relevance for CRP.

For outputs of convolutional layers, the relevance is summed across height and width channels for each feature.

See also See also IndexedConcepts.

ExplainableAI.IndexedConcepts — Type

IndexedConcepts(indices...)

Select concepts by indices for CRP.

For outputs of convolutional layers, the index refers to a feature dimension.

See also See also IndexedConcepts.

Index

ExplainableAI.AlphaBetaRule

ExplainableAI.CRP

ExplainableAI.ChainTuple

ExplainableAI.Composite

ExplainableAI.ConvLayer

ExplainableAI.DropoutLayer

ExplainableAI.EpsilonRule

ExplainableAI.Explanation

ExplainableAI.FirstLayerMap

ExplainableAI.FirstLayerTypeMap

ExplainableAI.FirstNTypeMap

ExplainableAI.FlatRule

ExplainableAI.GammaRule

ExplainableAI.GeneralizedGammaRule

ExplainableAI.GlobalMap

ExplainableAI.GlobalTypeMap

ExplainableAI.Gradient

ExplainableAI.IndexedConcepts

ExplainableAI.InputTimesGradient

ExplainableAI.InterpolationAugmentation

ExplainableAI.LRP

ExplainableAI.LastLayerMap

ExplainableAI.LastLayerTypeMap

ExplainableAI.LayerMap

ExplainableAI.NoiseAugmentation

ExplainableAI.NormalizationLayer

ExplainableAI.ParallelTuple

ExplainableAI.PassRule

ExplainableAI.PoolingLayer

ExplainableAI.RangeMap

ExplainableAI.RangeTypeMap

ExplainableAI.ReshapingLayer

ExplainableAI.TopNConcepts

ExplainableAI.WSquareRule

ExplainableAI.ZBoxRule

ExplainableAI.ZPlusRule

ExplainableAI.ZeroRule

ExplainableAI.EpsilonAlpha2Beta1

ExplainableAI.EpsilonAlpha2Beta1Flat

ExplainableAI.EpsilonGammaBox

ExplainableAI.EpsilonPlus

ExplainableAI.EpsilonPlusFlat

ExplainableAI.IntegratedGradients

ExplainableAI.LRP_CONFIG.supports_activation

ExplainableAI.LRP_CONFIG.supports_layer

ExplainableAI.SmoothGrad

ExplainableAI.analyze

ExplainableAI.canonize

ExplainableAI.flatten_model

ExplainableAI.heatmap

ExplainableAI.is_compatible

ExplainableAI.lrp_rules

ExplainableAI.modify_bias

ExplainableAI.modify_denominator

ExplainableAI.modify_input

ExplainableAI.modify_layer

ExplainableAI.modify_parameters

ExplainableAI.modify_weight

ExplainableAI.preprocess_imagenet

ExplainableAI.show_layer_indices

ExplainableAI.strip_softmax