Basic API

All methods in ExplainableAI.jl work by calling analyze on an input and an analyzer:

XAIBase.analyze — Function

analyze(input, method)
analyze(input, method, neuron_selection)

Apply the analyzer method for the given input, returning an Explanation. If neuron_selection is specified, e.g. the index of a specific output neuron, the explanation will be calculated for that neuron. Otherwise, the output neuron with the highest activation is automatically chosen.

See also Explanation and heatmap.

Keyword arguments

add_batch_dim: add batch dimension to the input without allocating. Default is false.

XAIBase.Explanation — Type

Return type of analyzers when calling analyze.

Fields

val: numerical output of the analyzer, e.g. an attribution or gradient

output: model output for the given analyzer input

output_selection: index of the output used for the explanation

analyzer: symbol corresponding the used analyzer, e.g. :Gradient or :LRP

heatmap: symbol indicating a preset heatmapping style, e.g. :attibution, :sensitivity or :cam

extras: optional named tuple that can be used by analyzers to return additional information.

XAIBase.heatmap — Function

heatmap(explanation)

Visualize Explanation from XAIBase as a vision heatmap. Assumes WHCN convention (width, height, channels, batchsize) for explanation.val.

Keyword arguments

colorscheme::Union{ColorScheme,Symbol}: color scheme from ColorSchemes.jl. Defaults to :seismic.

reduce::Symbol: Selects how color channels are reduced to a single number to apply a color scheme. The following methods can be selected, which are then applied over the color channels for each "pixel" in the array:

:sum: sum up color channels

:norm: compute 2-norm over the color channels

:maxabs: compute maximum(abs, x) over the color channels

Defaults to :sum.

rangescale::Symbol: Selects how the color channel reduced heatmap is normalized before the color scheme is applied. Can be either :extrema or :centered. Defaults to :centered.

process_batch::Bool: When heatmapping a batch, setting process_batch=true will apply the rangescale normalization to the entire batch instead of computing it individually for each sample in the batch. Defaults to false.

permute::Bool: Whether to flip W&H input channels. Default is true.

unpack_singleton::Bool: If false, heatmap will always return a vector of images. When heatmapping a batch with a single sample, setting unpack_singleton=true will unpack the singleton vector and directly return the image. Defaults to true.

heatmap(input, analyzer)

Compute an Explanation for a given input using the method analyzer and visualize it as a vision heatmap.

Any additional arguments and keyword arguments are passed to the analyzer. Refer to analyze for more information on available keyword arguments.

To customize the heatmapping style, first compute an explanation using analyze and then call heatmap on the explanation.

heatmap(explanation, text)

Visualize Explanation from XAIBase as text heatmap. Text should be a vector containing vectors of strings, one for each input in the batched explanation.

Keyword arguments

colorscheme::Union{ColorScheme,Symbol}: color scheme from ColorSchemes.jl. Defaults to :seismic.

rangescale::Symbol: selects how the color channel reduced heatmap is normalized before the color scheme is applied. Can be either :extrema or :centered. Defaults to :centered for use with the default color scheme :seismic.

Analyzers

ExplainableAI.Gradient — Type

Gradient(model)

Analyze model by calculating the gradient of a neuron activation with respect to the input.

ExplainableAI.InputTimesGradient — Type

InputTimesGradient(model)

Analyze model by calculating the gradient of a neuron activation with respect to the input. This gradient is then multiplied element-wise with the input.

ExplainableAI.SmoothGrad — Function

SmoothGrad(analyzer, [n=50, std=0.1, rng=GLOBAL_RNG])
SmoothGrad(analyzer, [n=50, distribution=Normal(0, σ²=0.01), rng=GLOBAL_RNG])

Analyze model by calculating a smoothed sensitivity map. This is done by averaging sensitivity maps of a Gradient analyzer over random samples in a neighborhood of the input, typically by adding Gaussian noise with mean 0.

References

Smilkov et al., SmoothGrad: removing noise by adding noise

ExplainableAI.IntegratedGradients — Function

IntegratedGradients(analyzer, [n=50])
IntegratedGradients(analyzer, [n=50])

Analyze model by using the Integrated Gradients method.

References

Sundararajan et al., Axiomatic Attribution for Deep Networks

Input augmentations

SmoothGrad and IntegratedGradients are special cases of the input augmentations NoiseAugmentation and InterpolationAugmentation, which can be applied as a wrapper to any analyzer:

ExplainableAI.NoiseAugmentation — Type

NoiseAugmentation(analyzer, n, [std=1, rng=GLOBAL_RNG])
NoiseAugmentation(analyzer, n, distribution, [rng=GLOBAL_RNG])

A wrapper around analyzers that augments the input with n samples of additive noise sampled from distribution. This input augmentation is then averaged to return an Explanation.

ExplainableAI.InterpolationAugmentation — Type

InterpolationAugmentation(model, [n=50])

A wrapper around analyzers that augments the input with n steps of linear interpolation between the input and a reference input (typically zero(input)). The gradients w.r.t. this augmented input are then averaged and multiplied with the difference between the input and the reference input.

Index

ExplainableAI.Gradient

ExplainableAI.InputTimesGradient

ExplainableAI.InterpolationAugmentation

ExplainableAI.NoiseAugmentation

XAIBase.Explanation

ExplainableAI.IntegratedGradients

ExplainableAI.SmoothGrad

XAIBase.analyze

XAIBase.heatmap