API Reference

All methods in ExplainabilityMethods.jl work by calling analyze on an input and an analyzer:

ExplainabilityMethods.analyze — Function

analyze(input, method)
analyze(input, method, neuron_selection)

Return raw classifier output and explanation. If neuron_selection is specified, the explanation will be calculated for that neuron. Otherwise, the output neuron with the highest activation is automatically chosen.

ExplainabilityMethods.heatmap — Function

heatmap(expl; kwargs...)

Visualize explanation.

Analyzers

ExplainabilityMethods.LRP — Type

LRP(c::Chain, r::AbstractLRPRule)
LRP(c::Chain, rs::AbstractVector{<:AbstractLRPRule})
LRP(layers::AbstractVector{LRPLayer})

Analyze model by applying Layer-Wise Relevance Propagation.

References

[1] G. Montavon et al., Layer-Wise Relevance Propagation: An Overview [2] W. Samek et al., Explaining Deep Neural Networks and Beyond: A Review of Methods and Applications

ExplainabilityMethods.Gradient — Type

Gradient(model)

Analyze model by calculating the gradient of a neuron activation with respect to the input.

ExplainabilityMethods.InputTimesGradient — Type

InputTimesGradient(model)

Analyze model by calculating the gradient of a neuron activation with respect to the input. This gradient is then multiplied element-wise with the input.

LRP

Rules

ExplainabilityMethods.ZeroRule — Type

ZeroRule()

Constructor for LRP-0 rule. Commonly used on upper layers.

ExplainabilityMethods.GammaRule — Type

GammaRule(; γ=0.25)

Constructor for LRP-$γ$ rule. Commonly used on lower layers.

Arguments:

γ: Optional multiplier for added positive weights, defaults to 0.25.

ExplainabilityMethods.EpsilonRule — Type

EpsilonRule(; ϵ=1f-6)

Constructor for LRP-$ϵ$ rule. Commonly used on middle layers.

Arguments:

ϵ: Optional stabilization parameter, defaults to 1f-6.

ExplainabilityMethods.ZBoxRule — Type

ZBoxRule()

Constructor for LRP-z^{\mathcal{B}}`-rule. Commonly used on the first layer for pixel input.

Custom rules

These utilities can be used to define custom rules without writing boilerplate code:

ExplainabilityMethods.modify_layer — Function

modify_layer(rule, layer)

Applies modify_params to layer if it has parameters

ExplainabilityMethods.modify_params — Function

modify_params!(rule, W, b)

Function that modifies weights and biases before applying relevance propagation.

ExplainabilityMethods.modify_denominator — Function

modify_denominator!(d, rule)

Function that modifies zₖ on the forward pass, e.g. for numerical stability.

Utilities

ExplainabilityMethods.strip_softmax — Function

strip_softmax(model)

Remove softmax activation on model output if it exists.

ExplainabilityMethods.flatten_chain — Function

flatten_chain(c)

Flatten a Flux chain containing Flux chains.

Index

ExplainabilityMethods.EpsilonRule

ExplainabilityMethods.GammaRule

ExplainabilityMethods.Gradient

ExplainabilityMethods.InputTimesGradient

ExplainabilityMethods.LRP

ExplainabilityMethods.ZBoxRule

ExplainabilityMethods.ZeroRule

ExplainabilityMethods.analyze

ExplainabilityMethods.flatten_chain

ExplainabilityMethods.heatmap

ExplainabilityMethods.modify_denominator

ExplainabilityMethods.modify_layer

ExplainabilityMethods.modify_params

ExplainabilityMethods.strip_softmax