Basics

All methods in ExplainableAI.jl work by calling analyze on an input and an analyzer:

ExplainableAI.analyze — Function

analyze(input, method)
analyze(input, method, neuron_selection)

Return an Explanation containing the attribution and the raw classifier output. If neuron_selection is specified, the explanation will be calculated for that neuron. Otherwise, the output neuron with the highest activation is automatically chosen.

Keyword arguments

add_batch_dim: add batch dimension to the input without allocating. Default is false.

ExplainableAI.Explanation — Type

Return type of analyzers when calling analyze.

Contains:

attribution: the analyzer's attribution

output: the model output

neuron_selection: the neuron index used for the attribution

analyzer: a symbol corresponding the used analyzer, e.g. :LRP

ExplainableAI.heatmap — Function

heatmap(explanation)
heatmap(input, analyzer)
heatmap(input, analyzer, neuron_selection)

Visualize explanation. Assumes Flux's WHCN convention (width, height, color channels, batch size).

Keyword arguments

cs::ColorScheme: ColorScheme that is applied. When calling heatmap with an Explanation or analyzer, the method default is selected. When calling heatmap with an array, the default is ColorSchemes.seismic.

reduce::Symbol: How the color channels are reduced to a single number to apply a colorscheme. The following methods can be selected, which are then applied over the color channels for each "pixel" in the attribution:

:sum: sum up color channels

:norm: compute 2-norm over the color channels

:maxabs: compute maximum(abs, x) over the color channels in

When calling heatmap with an Explanation or analyzer, the method default is selected. When calling heatmap with an array, the default is :sum.

rangescale::Symbol: How the color channel reduced heatmap is normalized before the colorscheme is applied. Can be either :extrema or :centered. When calling heatmap with an Explanation or analyzer, the method default is selected. When calling heatmap with an array, the default for use with the seismic colorscheme is :centered.

permute::Bool: Whether to flip W&H input channels. Default is true.

unpack_singleton::Bool: When heatmapping a batch with a single sample, setting unpack_singleton=true will return an image instead of an Vector containing a single image.

Note: keyword arguments can't be used when calling heatmap with an analyzer.

Analyzers

ExplainableAI.LRP — Type

LRP(model, rules)
LRP(model, composite)

Analyze model by applying Layer-Wise Relevance Propagation. The analyzer can either be created by passing an array of LRP-rules or by passing a composite, see Composite for an example.

Keyword arguments

skip_checks::Bool: Skip checks whether model is compatible with LRP and contains output softmax. Default is false.

verbose::Bool: Select whether the model checks should print a summary on failure. Default is true.

References

[1] G. Montavon et al., Layer-Wise Relevance Propagation: An Overview [2] W. Samek et al., Explaining Deep Neural Networks and Beyond: A Review of Methods and Applications

ExplainableAI.Gradient — Type

Gradient(model)

Analyze model by calculating the gradient of a neuron activation with respect to the input.

ExplainableAI.InputTimesGradient — Type

InputTimesGradient(model)

Analyze model by calculating the gradient of a neuron activation with respect to the input. This gradient is then multiplied element-wise with the input.

ExplainableAI.SmoothGrad — Function

SmoothGrad(analyzer, [n=50, std=0.1, rng=GLOBAL_RNG])
SmoothGrad(analyzer, [n=50, distribution=Normal(0, σ²=0.01), rng=GLOBAL_RNG])

Analyze model by calculating a smoothed sensitivity map. This is done by averaging sensitivity maps of a Gradient analyzer over random samples in a neighborhood of the input, typically by adding Gaussian noise with mean 0.

References

[1] Smilkov et al., SmoothGrad: removing noise by adding noise

ExplainableAI.IntegratedGradients — Function

IntegratedGradients(analyzer, [n=50])
IntegratedGradients(analyzer, [n=50])

Analyze model by using the Integrated Gradients method.

References

[1] Sundararajan et al., Axiomatic Attribution for Deep Networks

SmoothGrad and IntegratedGradients are special cases of the input augmentation wrappers NoiseAugmentation and InterpolationAugmentation, which can be applied as a wrapper to any analyzer:

ExplainableAI.NoiseAugmentation — Type

NoiseAugmentation(analyzer, n, [std=1, rng=GLOBAL_RNG])
NoiseAugmentation(analyzer, n, distribution, [rng=GLOBAL_RNG])

A wrapper around analyzers that augments the input with n samples of additive noise sampled from distribution. This input augmentation is then averaged to return an Explanation.

ExplainableAI.InterpolationAugmentation — Type

InterpolationAugmentation(model, [n=50])

A wrapper around analyzers that augments the input with n steps of linear interpolation between the input and a reference input (typically zero(input)). The gradients w.r.t. this augmented input are then averaged and multiplied with the difference between the input and the reference input.

Layer-wise Relevance Propagation

LRP rules

ExplainableAI.ZeroRule — Type

ZeroRule()

LRP-$0$ rule. Commonly used on upper layers.

Definition

Propagates relevance $R^{k+1}$ at layer output to $R^k$ at layer input according to

\[R_j^k = \sum_i \frac{w_{ij}a_j^k}{\sum_l w_{il}a_l^k+b_i} R_i^{k+1}\]

References

S. Bach et al., On Pixel-Wise Explanations for Non-Linear Classifier Decisions by Layer-Wise Relevance Propagation

ExplainableAI.EpsilonRule — Type

EpsilonRule([ϵ=1.0f-6])

LRP-$ϵ$ rule. Commonly used on middle layers.

Definition

Propagates relevance $R^{k+1}$ at layer output to $R^k$ at layer input according to

\[R_j^k = \sum_i\frac{w_{ij}a_j^k}{\epsilon +\sum_{l}w_{il}a_l^k+b_i} R_i^{k+1}\]

Optional arguments

ϵ: Optional stabilization parameter, defaults to 1f-6.

References

S. Bach et al., On Pixel-Wise Explanations for Non-Linear Classifier Decisions by Layer-Wise Relevance Propagation

ExplainableAI.GammaRule — Type

GammaRule([γ=0.25])

LRP-$γ$ rule. Commonly used on lower layers.

Definition

Propagates relevance $R^{k+1}$ at layer output to $R^k$ at layer input according to

\[R_j^k = \sum_i\frac{(w_{ij}+\gamma w_{ij}^+)a_j^k} {\sum_l(w_{il}+\gamma w_{il}^+)a_l^k+b_i} R_i^{k+1}\]

Optional arguments

γ: Optional multiplier for added positive weights, defaults to 0.25.

References

G. Montavon et al., Layer-Wise Relevance Propagation: An Overview

ExplainableAI.WSquareRule — Type

WSquareRule()

LRP-$w²$ rule. Commonly used on the first layer when values are unbounded.

Definition

Propagates relevance $R^{k+1}$ at layer output to $R^k$ at layer input according to

\[R_j^k = \sum_i\frac{w_{ij}^2}{\sum_l w_{il}^2} R_i^{k+1}\]

References

G. Montavon et al., Explaining Nonlinear Classification Decisions with Deep Taylor Decomposition

ExplainableAI.FlatRule — Type

FlatRule()

LRP-Flat rule. Similar to the WSquareRule, but with all weights set to one and all bias terms set to zero.

Definition

Propagates relevance $R^{k+1}$ at layer output to $R^k$ at layer input according to

\[R_j^k = \sum_i\frac{1}{\sum_l 1} R_i^{k+1} = \frac{1}{n}\sum_i R_i^{k+1}\]

where $n$ is the number of input neurons connected to the output neuron at index $i$.

References

S. Lapuschkin et al., Unmasking Clever Hans predictors and assessing what machines really learn

ExplainableAI.AlphaBetaRule — Type

AlphaBetaRule([α=2.0], [β=1.0])

LRP-$αβ$ rule. Weights positive and negative contributions according to the parameters α and β respectively. The difference α-β must be equal to one. Commonly used on lower layers.

Definition

Propagates relevance $R^{k+1}$ at layer output to $R^k$ at layer input according to

\[R_j^k = \sum_i\left( \alpha\frac{\left(w_{ij}a_j^k\right)^+}{\sum_l\left(w_{il}a_l^k+b_i\right)^+} -\beta\frac{\left(w_{ij}a_j^k\right)^-}{\sum_l\left(w_{il}a_l^k+b_i\right)^-} \right) R_i^{k+1}\]

Optional arguments

α: Multiplier for the positive output term, defaults to 2.0.

β: Multiplier for the negative output term, defaults to 1.0.

References

S. Bach et al., On Pixel-Wise Explanations for Non-Linear Classifier Decisions by Layer-Wise Relevance Propagation

G. Montavon et al., Layer-Wise Relevance Propagation: An Overview

ExplainableAI.ZPlusRule — Type

ZPlusRule()

LRP-$z⁺$ rule. Commonly used on lower layers.

Equivalent to AlphaBetaRule(1.0f0, 0.0f0), but slightly faster. See also AlphaBetaRule.

Definition

Propagates relevance $R^{k+1}$ at layer output to $R^k$ at layer input according to

\[R_j^k = \sum_i\frac{\left(w_{ij}a_j^k\right)^+}{\sum_l\left(w_{il}a_l^k+b_i\right)^+} R_i^{k+1}\]

References

S. Bach et al., On Pixel-Wise Explanations for Non-Linear Classifier Decisions by Layer-Wise Relevance Propagation

G. Montavon et al., Explaining Nonlinear Classification Decisions with Deep Taylor Decomposition

ExplainableAI.ZBoxRule — Type

ZBoxRule(low, high)

LRP-$zᴮ$-rule. Commonly used on the first layer for pixel input.

The parameters low and high should be set to the lower and upper bounds of the input features, e.g. 0.0 and 1.0 for raw image data. It is also possible to provide two arrays of that match the input size.

Definition

Propagates relevance $R^{k+1}$ at layer output to $R^k$ at layer input according to

\[R_j^k=\sum_i \frac{w_{ij}a_j^k - w_{ij}^{+}l_j - w_{ij}^{-}h_j} {\sum_l w_{il}a_l^k+b_i - \left(w_{il}^{+}l_l+b_i^{+}\right) - \left(w_{il}^{-}h_l+b_i^{-}\right)} R_i^{k+1}\]

References

G. Montavon et al., Layer-Wise Relevance Propagation: An Overview

ExplainableAI.PassRule — Type

PassRule()

Pass-through rule. Passes relevance through to the lower layer.

Supports layers with constant input and output shapes, e.g. reshaping layers.

Definition

Propagates relevance $R^{k+1}$ at layer output to $R^k$ at layer input according to

\[R_j^k = R_j^{k+1}\]

Custom rules

These utilities can be used to define custom rules without writing boilerplate code:

ExplainableAI.modify_input — Function

modify_input(rule, input)

Modify input activation before computing relevance propagation.

ExplainableAI.modify_denominator — Function

modify_denominator(rule, d)

Modify denominator $z$ for numerical stability on the forward pass.

ExplainableAI.modify_param! — Function

modify_param!(rule, W)
modify_param!(rule, b)

Inplace-modify parameters before computing the relevance.

ExplainableAI.modify_layer! — Function

modify_layer!(rule, layer; ignore_bias=false)

In-place modify layer parameters by calling modify_param! before computing relevance propagation.

Note

When implementing a custom modify_layer! function, modify_param! will not be called.

ExplainableAI.check_compat — Function

check_compat(rule, layer)

Check compatibility of a LRP-Rule with layer type.

Note

When implementing a custom check_compat function, return nothing if checks passed, otherwise throw an ArgumentError.

ExplainableAI.LRP_CONFIG.supports_layer — Function

LRP_CONFIG.supports_layer(layer)

Check whether LRP can be used on a layer or a Chain. To extend LRP to your own layers, define:

LRP_CONFIG.supports_layer(::MyLayer) = true          # for structs
LRP_CONFIG.supports_layer(::typeof(mylayer)) = true  # for functions

ExplainableAI.LRP_CONFIG.supports_activation — Function

LRP_CONFIG.supports_activation(σ)

Check whether LRP can be used on a given activation function. To extend LRP to your own activation functions, define:

LRP_CONFIG.supports_activation(::typeof(myactivation)) = true  # for functions
LRP_CONFIG.supports_activation(::MyActivation) = true          # for structs

Composites

ExplainableAI.Composite — Type

Composite([default_rule=LRPZero()], primitives...)

Automatically contructs a list of LRP-rules by sequentially applying composite primitives.

Primitives

To apply a single rule, use:

LayerRule to apply a rule to the n-th layer of a model

GlobalRule to apply a rule to all layers

RangeRule to apply a rule to a positional range of layers

FirstLayerRule to apply a rule to the first layer

LastLayerRule to apply a rule to the last layer

To apply a set of rules to layers based on their type, use:

GlobalTypeRule to apply a dictionary that maps layer types to LRP-rules

RangeTypeRule for a TypeRule on generalized ranges

FirstLayerTypeRule for a TypeRule on the first layer of a model

LastLayerTypeRule for a TypeRule on the last layer

FirstNTypeRule for a TypeRule on the first n layers

LastNTypeRule for a TypeRule on the last n layers

Example

Using a flattened VGG11 model:

julia> composite = Composite(
           GlobalTypeRule(
               ConvLayer => AlphaBetaRule(),
               Dense => EpsilonRule(),
               PoolingLayer => EpsilonRule(),
               DropoutLayer => PassRule(),
               ReshapingLayer => PassRule(),
           ),
           FirstNTypeRule(7, Conv => FlatRule()),
       );

julia> analyzer = LRP(model, composite);

julia> analyzer.rules
19-element Vector{AbstractLRPRule}:
 FlatRule()
 EpsilonRule{Float32}(1.0f-6)
 FlatRule()
 EpsilonRule{Float32}(1.0f-6)
 FlatRule()
 FlatRule()
 EpsilonRule{Float32}(1.0f-6)
 AlphaBetaRule{Float32}(2.0f0, 1.0f0)
 AlphaBetaRule{Float32}(2.0f0, 1.0f0)
 EpsilonRule{Float32}(1.0f-6)
 AlphaBetaRule{Float32}(2.0f0, 1.0f0)
 AlphaBetaRule{Float32}(2.0f0, 1.0f0)
 EpsilonRule{Float32}(1.0f-6)
 PassRule()
 EpsilonRule{Float32}(1.0f-6)
 PassRule()
 EpsilonRule{Float32}(1.0f-6)
 PassRule()
 EpsilonRule{Float32}(1.0f-6)

Composite primitives

Composite primitives that apply a single rule:

ExplainableAI.LayerRule — Type

LayerRule(n, rule)

Composite primitive that applies LRP-rule rule to the n-th layer in the model.

See Composite for an example.

ExplainableAI.GlobalRule — Type

GlobalRule(rule)

Composite primitive that applies LRP-rule rule to all layers in the model.

See Composite for an example.

ExplainableAI.RangeRule — Type

RangeRule(range, rule)

Composite primitive that applies LRP-rule rule to the specified positional range of layers in the model.

See Composite for an example.

ExplainableAI.FirstLayerRule — Type

FirstLayerRule(rule)

Composite primitive that applies LRP-rule rule to the first layer in the model.

See Composite for an example.

ExplainableAI.LastLayerRule — Type

LastLayerRule(rule)

Composite primitive that applies LRP-rule rule to the last layer in the model.

See Composite for an example.

Composite primitives that apply a set of rules to multiple layers:

ExplainableAI.GlobalTypeRule — Type

GlobalTypeRule(map)

Composite primitive that maps layer types to LRP rules based on a list of type-rule-pairs map.

See Composite for an example.

ExplainableAI.RangeTypeRule — Type

RangeTypeRule(range, map)

Composite primitive that maps layer types to LRP rules based on a list of type-rule-pairs map within the specified range of layers in the model.

See Composite for an example.

ExplainableAI.FirstLayerTypeRule — Type

FirstLayerTypeRule(map)

Composite primitive that maps the type of the first layer of the model to LRP rules based on a list of type-rule-pairs map.

See Composite for an example.

ExplainableAI.LastLayerTypeRule — Type

LastLayerTypeRule(map)

Composite primitive that maps the type of the last layer of the model to LRP rules based on a list of type-rule-pairs map.

See Composite for an example.

ExplainableAI.FirstNTypeRule — Type

FirstNTypeRule(n, map)

Composite primitive that maps layer types to LRP rules based on a list of type-rule-pairs map within the first n layers in the model.

See Composite for an example.

ExplainableAI.LastNTypeRule — Type

LastNTypeRule(n, map)

Composite primitive that maps layer types to LRP rules based on a list of type-rule-pairs map within the last n layers in the model.

See Composite for an example.

Default composites

ExplainableAI.EpsilonGammaBox — Function

EpsilonGammaBox(low, high; [epsilon=1.0f-6, gamma=0.25f0])

Composite using the following primitives:

julia> EpsilonGammaBox(-3.0f0, 3.0f0)
Composite(
  GlobalTypeRule(     # on all layers
    Union{Flux.Conv, Flux.ConvTranspose, Flux.CrossCor} => ExplainableAI.GammaRule{Float32}(0.25f0),
    Flux.Dense => ExplainableAI.EpsilonRule{Float32}(1.0f-6),
    Union{typeof(Flux.dropout), Flux.AlphaDropout, Flux.Dropout} => ExplainableAI.PassRule(),
    Union{typeof(Flux.flatten), typeof(MLUtils.flatten)} => ExplainableAI.PassRule(),
  ),
  FirstLayerTypeRule( # on first layer
    Union{Flux.Conv, Flux.ConvTranspose, Flux.CrossCor} => ExplainableAI.ZBoxRule{Float32}(-3.0f0, 3.0f0),
  ),
)

ExplainableAI.EpsilonPlus — Function

EpsilonPlus(; [epsilon=1.0f-6])

Composite using the following primitives:

julia> EpsilonPlus()
Composite(
  GlobalTypeRule(     # on all layers
    Union{Flux.Conv, Flux.ConvTranspose, Flux.CrossCor} => ExplainableAI.ZPlusRule(),
    Flux.Dense => ExplainableAI.EpsilonRule{Float32}(1.0f-6),
    Union{typeof(Flux.dropout), Flux.AlphaDropout, Flux.Dropout} => ExplainableAI.PassRule(),
    Union{typeof(Flux.flatten), typeof(MLUtils.flatten)} => ExplainableAI.PassRule(),
  ),
)

ExplainableAI.EpsilonAlpha2Beta1 — Function

EpsilonAlpha2Beta1(; [epsilon=1.0f-6])

Composite using the following primitives:

julia> EpsilonAlpha2Beta1()
Composite(
  GlobalTypeRule(     # on all layers
    Union{Flux.Conv, Flux.ConvTranspose, Flux.CrossCor} => ExplainableAI.AlphaBetaRule{Float32}(2.0f0, 1.0f0),
    Flux.Dense => ExplainableAI.EpsilonRule{Float32}(1.0f-6),
    Union{typeof(Flux.dropout), Flux.AlphaDropout, Flux.Dropout} => ExplainableAI.PassRule(),
    Union{typeof(Flux.flatten), typeof(MLUtils.flatten)} => ExplainableAI.PassRule(),
  ),
)

ExplainableAI.EpsilonPlusFlat — Function

EpsilonPlusFlat(; [epsilon=1.0f-6])

Composite using the following primitives:

julia> EpsilonPlusFlat()
Composite(
  GlobalTypeRule(     # on all layers
    Union{Flux.Conv, Flux.ConvTranspose, Flux.CrossCor} => ExplainableAI.ZPlusRule(),
    Flux.Dense => ExplainableAI.EpsilonRule{Float32}(1.0f-6),
    Union{typeof(Flux.dropout), Flux.AlphaDropout, Flux.Dropout} => ExplainableAI.PassRule(),
    Union{typeof(Flux.flatten), typeof(MLUtils.flatten)} => ExplainableAI.PassRule(),
  ),
  FirstLayerTypeRule( # on first layer
    Union{Flux.Conv, Flux.ConvTranspose, Flux.CrossCor} => ExplainableAI.FlatRule(),
    Flux.Dense => ExplainableAI.FlatRule(),
  ),
)

ExplainableAI.EpsilonAlpha2Beta1Flat — Function

EpsilonAlpha2Beta1Flat(; [epsilon=1.0f-6])

Composite using the following primitives:

julia> EpsilonAlpha2Beta1Flat()
Composite(
  GlobalTypeRule(     # on all layers
    Union{Flux.Conv, Flux.ConvTranspose, Flux.CrossCor} => ExplainableAI.AlphaBetaRule{Float32}(2.0f0, 1.0f0),
    Flux.Dense => ExplainableAI.EpsilonRule{Float32}(1.0f-6),
    Union{typeof(Flux.dropout), Flux.AlphaDropout, Flux.Dropout} => ExplainableAI.PassRule(),
    Union{typeof(Flux.flatten), typeof(MLUtils.flatten)} => ExplainableAI.PassRule(),
  ),
  FirstLayerTypeRule( # on first layer
    Union{Flux.Conv, Flux.ConvTranspose, Flux.CrossCor} => ExplainableAI.FlatRule(),
    Flux.Dense => ExplainableAI.FlatRule(),
  ),
)

Utilities

ExplainableAI.strip_softmax — Function

strip_softmax(model)

Remove softmax activation on model output if it exists.

ExplainableAI.flatten_model — Function

flatten_model(c)

Flatten a Flux chain containing Flux chains.

ExplainableAI.canonize — Function

canonize(model)

Canonize model by flattening it and fusing BatchNorm layers into preceding Dense and Conv layers with linear activation functions.

Index

ExplainableAI.AlphaBetaRule

ExplainableAI.Composite

ExplainableAI.EpsilonRule

ExplainableAI.Explanation

ExplainableAI.FirstLayerRule

ExplainableAI.FirstLayerTypeRule

ExplainableAI.FirstNTypeRule

ExplainableAI.FlatRule

ExplainableAI.GammaRule

ExplainableAI.GlobalRule

ExplainableAI.GlobalTypeRule

ExplainableAI.Gradient

ExplainableAI.InputTimesGradient

ExplainableAI.InterpolationAugmentation

ExplainableAI.LRP

ExplainableAI.LastLayerRule

ExplainableAI.LastLayerTypeRule

ExplainableAI.LastNTypeRule

ExplainableAI.LayerRule

ExplainableAI.NoiseAugmentation

ExplainableAI.PassRule

ExplainableAI.RangeRule

ExplainableAI.RangeTypeRule

ExplainableAI.WSquareRule

ExplainableAI.ZBoxRule

ExplainableAI.ZPlusRule

ExplainableAI.ZeroRule

ExplainableAI.EpsilonAlpha2Beta1

ExplainableAI.EpsilonAlpha2Beta1Flat

ExplainableAI.EpsilonGammaBox

ExplainableAI.EpsilonPlus

ExplainableAI.EpsilonPlusFlat

ExplainableAI.IntegratedGradients

ExplainableAI.LRP_CONFIG.supports_activation

ExplainableAI.LRP_CONFIG.supports_layer

ExplainableAI.SmoothGrad

ExplainableAI.analyze

ExplainableAI.canonize

ExplainableAI.check_compat

ExplainableAI.flatten_model

ExplainableAI.heatmap

ExplainableAI.modify_denominator

ExplainableAI.modify_input

ExplainableAI.modify_layer!

ExplainableAI.modify_param!

ExplainableAI.strip_softmax