Basics

All methods in ExplainableAI.jl work by calling analyze on an input and an analyzer:

ExplainableAI.analyze — Function

analyze(input, method)
analyze(input, method, neuron_selection)

Return raw classifier output and explanation. If neuron_selection is specified, the explanation will be calculated for that neuron. Otherwise, the output neuron with the highest activation is automatically chosen.

Keyword arguments

add_batch_dim: add batch dimension to the input without allocating. Default is false.

ExplainableAI.heatmap — Function

heatmap(expl::Explanation; kwargs...)
heatmap(attr::AbstractArray; kwargs...)

heatmap(input, analyzer::AbstractXAIMethod)
heatmap(input, analyzer::AbstractXAIMethod, neuron_selection::Int)

Visualize explanation. Assumes Flux's WHCN convention (width, height, color channels, batch size).

Keyword arguments

cs::ColorScheme: ColorScheme that is applied. When calling heatmap with an Explanation or analyzer, the method default is selected. When calling heatmap with an array, the default is ColorSchemes.seismic.

reduce::Symbol: How the color channels are reduced to a single number to apply a colorscheme. The following methods can be selected, which are then applied over the color channels for each "pixel" in the attribution:

:sum: sum up color channels

:norm: compute 2-norm over the color channels

:maxabs: compute maximum(abs, x) over the color channels in

When calling heatmap with an Explanation or analyzer, the method default is selected. When calling heatmap with an array, the default is :sum.

rangescale::Symbol: How the color channel reduced heatmap is normalized before the colorscheme is applied. Can be either :extrema or :centered. When calling heatmap with an Explanation or analyzer, the method default is selected. When calling heatmap with an array, the default for use with the seismic colorscheme is :centered.

permute::Bool: Whether to flip W&H input channels. Default is true.

unpack_singleton::Bool: When heatmapping a batch with a single sample, setting unpack_singleton=true will return an image instead of an Vector containing a single image.

Note: these keyword arguments can't be used when calling heatmap with an analyzer.

Analyzers

ExplainableAI.LRP — Type

LRP(c::Chain, r::AbstractLRPRule)
LRP(c::Chain, rs::AbstractVector{<:AbstractLRPRule})

Analyze model by applying Layer-Wise Relevance Propagation.

Keyword arguments

skip_checks::Bool: Skip checks whether model is compatible with LRP and contains output softmax. Default is false.

verbose::Bool: Select whether the model checks should print a summary on failure. Default is true.

References

[1] G. Montavon et al., Layer-Wise Relevance Propagation: An Overview [2] W. Samek et al., Explaining Deep Neural Networks and Beyond: A Review of Methods and Applications

ExplainableAI.Gradient — Type

Gradient(model)

Analyze model by calculating the gradient of a neuron activation with respect to the input.

ExplainableAI.InputTimesGradient — Type

InputTimesGradient(model)

Analyze model by calculating the gradient of a neuron activation with respect to the input. This gradient is then multiplied element-wise with the input.

ExplainableAI.SmoothGrad — Function

SmoothGrad(analyzer, [n=50, std=0.1, rng=GLOBAL_RNG])
SmoothGrad(analyzer, [n=50, distribution=Normal(0, σ²=0.01), rng=GLOBAL_RNG])

Analyze model by calculating a smoothed sensitivity map. This is done by averaging sensitivity maps of a Gradient analyzer over random samples in a neighborhood of the input, typically by adding Gaussian noise with mean 0.

References

[1] Smilkov et al., SmoothGrad: removing noise by adding noise

ExplainableAI.IntegratedGradients — Function

IntegratedGradients(analyzer, [n=50])
IntegratedGradients(analyzer, [n=50])

Analyze model by using the Integrated Gradients method.

References

[1] Sundararajan et al., Axiomatic Attribution for Deep Networks

SmoothGrad and IntegratedGradients are special cases of the input augmentation wrappers NoiseAugmentation and InterpolationAugmentation, which can be applied as a wrapper to any analyzer:

ExplainableAI.NoiseAugmentation — Type

NoiseAugmentation(analyzer, n, [std=1, rng=GLOBAL_RNG])
NoiseAugmentation(analyzer, n, distribution, [rng=GLOBAL_RNG])

A wrapper around analyzers that augments the input with n samples of additive noise sampled from distribution. This input augmentation is then averaged to return an Explanation.

ExplainableAI.InterpolationAugmentation — Type

InterpolationAugmentation(model, [n=50])

A wrapper around analyzers that augments the input with n steps of linear interpolation between the input and a reference input (typically zero(input)). The gradients w.r.t. this augmented input are then averaged and multiplied with the difference between the input and the reference input.

LRP

Rules

ExplainableAI.ZeroRule — Type

ZeroRule()

LRP-0 rule. Commonly used on upper layers.

References

[1]: S. Bach et al., On Pixel-Wise Explanations for Non-Linear Classifier Decisions by Layer-Wise Relevance Propagation

ExplainableAI.EpsilonRule — Type

EpsilonRule([ϵ=1.0f-6])

LRP-$ϵ$ rule. Commonly used on middle layers.

Arguments:

ϵ: Optional stabilization parameter, defaults to 1f-6.

References

[1]: S. Bach et al., On Pixel-Wise Explanations for Non-Linear Classifier Decisions by Layer-Wise Relevance Propagation

ExplainableAI.GammaRule — Type

GammaRule([γ=0.25])

LRP-$γ$ rule. Commonly used on lower layers.

Arguments:

γ: Optional multiplier for added positive weights, defaults to 0.25.

References

[1]: G. Montavon et al., Layer-Wise Relevance Propagation: An Overview

ExplainableAI.WSquareRule — Type

WSquareRule()

LRP-$W^2$ rule. Commonly used on the first layer when values are unbounded.

References

[1]: G. Montavon et al., Explaining nonlinear classification decisions with deep Taylor decomposition

ExplainableAI.AlphaBetaRule — Type

AlphaBetaRule(alpha, beta)
AlphaBetaRule([alpha=2.0], [beta=1.0])

LRP-$lphaeta$ rule. Weights positive and negative contributions according to the parameters alpha and beta respectively. The difference alpha - beta must be equal one. Commonly used on lower layers.

Arguments:

alpha: Multiplier for the positive output term, defaults to 2.0.

beta: Multiplier for the negative output term, defaults to 1.0.

References

[1]: S. Bach et al., On Pixel-Wise Explanations for Non-Linear Classifier Decisions by Layer-Wise Relevance Propagation [2]: G. Montavon et al., Layer-Wise Relevance Propagation: An Overview

ExplainableAI.FlatRule — Type

FlatRule()

LRP-Flat rule. Similar to the WSquareRule, but with all parameters set to one.

References

[1]: S. Lapuschkin et al., Unmasking Clever Hans predictors and assessing what machines really learn

ExplainableAI.ZBoxRule — Type

ZBoxRule(low, high)

LRP-$z^{\mathcal{B}}$-rule. Commonly used on the first layer for pixel input.

The parameters low and high should be set to the lower and upper bounds of the input features, e.g. 0.0 and 1.0 for raw image data. It is also possible to provide two arrays of that match the input size.

References

[1]: G. Montavon et al., Explaining nonlinear classification decisions with deep Taylor decomposition

ExplainableAI.PassRule — Type

PassRule()

Pass-through rule. Passes relevance through to the lower layer. Supports reshaping layers.

Custom rules

These utilities can be used to define custom rules without writing boilerplate code:

ExplainableAI.modify_input — Function

modify_input(rule, input)

Modify input activation before computing relevance propagation.

ExplainableAI.modify_denominator — Function

modify_denominator(rule, d)

Modify denominator $z$ for numerical stability on the forward pass.

ExplainableAI.modify_param! — Function

modify_param!(rule, W)
modify_param!(rule, b)

Inplace-modify parameters before computing the relevance.

ExplainableAI.modify_layer! — Function

modify_layer!(rule, layer; ignore_bias=false)

In-place modify layer parameters by calling modify_param! before computing relevance propagation.

Note

When implementing a custom modify_layer! function, modify_param! will not be called.

Missing docstring.

Missing docstring for check_compat. Check Documenter's build log for details.

ExplainableAI.LRP_CONFIG.supports_layer — Function

LRP_CONFIG.supports_layer(layer)

Check whether LRP can be used on a layer or a Chain. To extend LRP to your own layers, define:

LRP_CONFIG.supports_layer(::MyLayer) = true          # for structs
LRP_CONFIG.supports_layer(::typeof(mylayer)) = true  # for functions

ExplainableAI.LRP_CONFIG.supports_activation — Function

LRP_CONFIG.supports_activation(σ)

Check whether LRP can be used on a given activation function. To extend LRP to your own activation functions, define:

LRP_CONFIG.supports_activation(::typeof(myactivation)) = true  # for functions
LRP_CONFIG.supports_activation(::MyActivation) = true          # for structs

Utilities

ExplainableAI.strip_softmax — Function

strip_softmax(model)

Remove softmax activation on model output if it exists.

ExplainableAI.flatten_model — Function

flatten_model(c)

Flatten a Flux chain containing Flux chains.

ExplainableAI.canonize — Function

canonize(model)

Canonize model by flattening it and fusing BatchNorm layers into preceding Dense and Conv layers with linear activation functions.

Index

ExplainableAI.AlphaBetaRule

ExplainableAI.EpsilonRule

ExplainableAI.FlatRule

ExplainableAI.GammaRule

ExplainableAI.Gradient

ExplainableAI.InputTimesGradient

ExplainableAI.InterpolationAugmentation

ExplainableAI.LRP

ExplainableAI.NoiseAugmentation

ExplainableAI.PassRule

ExplainableAI.WSquareRule

ExplainableAI.ZBoxRule

ExplainableAI.ZeroRule

ExplainableAI.IntegratedGradients

ExplainableAI.LRP_CONFIG.supports_activation

ExplainableAI.LRP_CONFIG.supports_layer

ExplainableAI.SmoothGrad

ExplainableAI.analyze

ExplainableAI.canonize

ExplainableAI.flatten_model

ExplainableAI.heatmap

ExplainableAI.modify_denominator

ExplainableAI.modify_input

ExplainableAI.modify_layer!

ExplainableAI.modify_param!

ExplainableAI.strip_softmax